Move code to better files.
